requirements.txt
albumentations==1.1.0
Flask==2.0.3
iglovikov_helper_functions==0.0.53
matplotlib==3.5.1
numpy==1.22.1
opencv_python_headless==4.5.5.64
pandas==1.3.5
Pillow==9.1.0
scikit_learn==1.0.2
segmentation_models_pytorch==0.2.1
tensorflow==2.8.0
torch==1.11.0